Quick and dirty way to make the Kinect track a foot in vr
Right now the tracker is behind the player by, like a meter-ish, so I'll
need to consider a way to set the proper position relative to the
players head, maybe also using the kinect skeleton.
Also, the code at this point is disgusting because I coded from 10pm-
1:30am. Currently working on a refactor for it so that it is not so
goddamn bad. (Removing global variables and replacing them with
referenced functions and decoupling the draw and proccessing stuff on
the skeleton side)