Added in a stopgap to fix controllers not working bug
Currently, the controllers are reconnected every frame, as before they
weren't due to some weird bug. I'll need to investigate this to remove
this eventually. However, from testing there hasn't been any noticeable
or even observable increase in processing usage. But this still needs to
be investigated as something's fucky