Stable: Moved positionOffset and zeroed into KinectHandler
One step closer to a world without globals. As a consequence though,
device now needs to be passed the offset in the update function. Also I
have an idea: instead of repeating the basic draw skeleton for kinectv1
and two, just have them in the interface, and redefine DrawBone and
SkeleToScreen for each implementation