v2 Rot Smoothing Fixed, and adjusted UI buttons
The problem was due to the length of the quaternion not being sqrt'd, which would cause massive innaccuracy to joints *occasionally*.  Clamp function is currently unused, but may be needed for eventual development of the smoothing. It currently works better than no smoothing,  but I need to adjust the queue - as right now only one extra result is used, meaning that it's not smoothing as much as it could.

Additionally I adjusted the GUI buttons per PanDahlia's feedback to be checkboxes, and they now work much more intuitively.