Unicode Config Loading
The sfml font loading stuff doesn't need an absolute path, it can just use a relative one, but as far as I've found, the config needs to loaded with an absolute path. Before, Unicode symbols, basically anything that's not plain ASCII english wouldn't be in the string, and thus the config wouldn't load properly. This has been fixed by changing the relevant strings to wstrings.