Trackers can now be selected, spawned and shown within VR for PSMoves
Each temp tracker in the GUI selection also stores the position tracking option (IMU, Skele etc.) so that the tracking methods can quickly skip anything it's not supposed to process.
Also, the trackingPool is now updated from the PSMoves, and a missing reference variable has been added to the IMU methods, as previously they only updated the copy.