Removed redundant code in VrController, changed Vr Scene Type
In the new openvr sdk, it states in the wiki that in utility mode "Only the IVRSettings and IVRApplications interfaces are guaranteed to work. This application type is appropriate for things like installers."
Which does not include the input system, or just about anything else. So to be safe, going to have to change the type. K2VR needs SteamVR open anyway