Kinect V1 and V2 have correct foot rotations
Previously, the bone was set to the correct one for the V2, but the V1 was actually reversed, so in the main process loop of each process, they set their respective bone types in the global KinectSettings for the bones with the correct rotations.

Added an invalid option for the joint types to specify if the device is using the joints or not.

Also added warning to tracker spawn button, as it seems that the glitch where your controllers don't work after spawning the trackers occurs less if the headset is on and being used.