Fixed Scaling, added Font Size scalebox
Now, the window can be resized as much as you want, and the font size increased and saved in a config file.

More work needs to be done to find a more performant solution, and adjusting the sizes of the windows by default - but Win 10 doesn't let me test resolutions that my monitor doesn't support (but Win 7 did?). So I'm kind of relying on other people right now.