Added logging out the wazoo, fixed potential memory leaks in color and depth frames
From now on all logging will be done with easylogging. The changes to the .cc and .h includes are made because the MSVC compiler seems to not want to use the compiler option for the log file name, but it works when added directly to the file.

Also, if a frame wasn't received for the color or depth sensors on the kinect V2, it would constantly leak memory. This has now been fixed.