Fixed redundant log messages, and trackers not being properly updated
The V2 kinect should now avoid posting the 'pending' (waiting for frame) logs when it can't get the frame - as that simply means it's relying solely on smoothing for the current frame.

The main kinect status label was also constantly spitting out 'status changed' to the current status, which has now been fixed.

Additionally, the trackers now correctly update in their default updateTrackersWithSkeletonPosition function, as previously the lack of a reference variable meant that any local changes applied to the device were really being applied to the foreach variable only.