PSMove's mostly work, auto-calibration process nearly spoton
Also, changed behaviour of devices to use the full pose provided by a device handler in case both the position and rotation deviceId's are the same. In the future though, any offset specific stuff should also be directly added to the position itself, so that it's ready for any tracker to use without the full pose (i.e. device driver, head, world offsets)