SteamVR Devices now registered as devices in K2VR
Now, the position or rotation of any steamvr device (sensors excluded) can be used on any device. Essentially, this paves the way for easier integration of the Virtual Hips setting.

Additionally, commented out the old deprecated VRController updating, as it was using CPU time for nothing.

Also made the selection of the trackers recognise whether the kinect joints are spawned into the box or not, and adjust the value accordingly - as with Kinectless, the joints are no longer added because they don't exist