Secondary Offset Calibration
This is intended to fix the issue people had where their devices would always be in the wrong position.

Add any device to the list, spawn it, and in the Adv. Trackers tab the Calibrate VR Offset button should appear. Click it while in VR, and the offsets should hopefully update properly