Added ability to spawn PSEyes representations in VR
Stilll needs more work, to make sure they don't register as actual trackers, like the kinect arrow. But now, there's a way to actually see just how far off the cameras are in order to better diagnose what the issues are with the current coordinate system.

Also added (and left disabled) a way to determine if the PSMove scale is off, and adjust the scale factor by that amount.