Quick and dirty way to get oneclick tracker reinitialisation across sessions
This will be cleaned up later, but after you spawn the trackers the regular way it'll save to a lastTrackers.cfg, and then you can click the 'Spawn same trackers as last session' button right under the regular one, and it should hopefully work for all device types.

It will likely crash when the IDs get mixed up, so that's first priority for the next commit.