Fixed PSMoveHandler not starting properly automatically, and fixed important tracker properties not being resaved after respawning them
Now, each device is identified by it's serial number too. If there's a mismatch in the case of the ID's, K2VR will try it's best to find it, but if it can't it'll stop adding the trackers to the list so that they can be closely investigated without crashing.

Also, PSMoveHandler wasn't being started correctly because it didn't get added to the vector of handlers in order to actually get updated. There's probably a better solution for this, but hey, it works.

PSMoves are identified by their device serial

PSEyes are identified by their device path