Fixed numerous crashes in Adv Tracker tab
Now, pressing Add Hand Controllers should no longer cause an index error and raise an exception.
Similarly, pressing beep or boop to identify a device which doesn't have a proper implementation will no longer crash either.
Adding a tracker to the list when nothing is selected will not crash, now.

In each of these cases, the adding the tracker to the list will abort, and nothing will actually be put into the list to spawn.

Also, fixed a crash when the kinect v1 sensor is a nullptr (caused crashes preventing further diagnosis)