HUGE: Migration of Kinect Tracking to new system
Optimisation still needs to follow, but as is it performs about the same as the old one - while being a hell of a lot more flexible.

Hopefully, this means that the changes made in regards to the additional offsets for some users (Particularly with the Vive) and the doubled Arrow rotation values should be fixed.

Impending 0.5.11 release incoming.