Nice little hack to fall back to legacy input when Valve breaks things
Additionally, changed the way input is handled, and the Calibrator no longer updates the data (Makes sense in hindsight) and the hack works by updating what is supposed to be the SteamVR obtained Input data with direct bitmasks of the controller data through the deprecated system. Should hold through for any broken stuff.